Move NodeSharingJenkinsRule to utils package
