+++ jth-tests/src/test/java/com/redhat/jenkins/nodesharing/PoolTest.java
@@ -24,9 +24,9 @@
 package com.redhat.jenkins.nodesharing;
 
 import com.cloudbees.plugins.credentials.common.UsernamePasswordCredentials;
-import com.redhat.jenkins.nodesharing.NodeSharingJenkinsRule.BlockingTask;
-import com.redhat.jenkins.nodesharing.NodeSharingJenkinsRule.MockTask;
-import com.redhat.jenkins.nodesharing.utils.BlockingBuilder;
+import com.redhat.jenkins.nodesharing.utils.NodeSharingJenkinsRule;
+import com.redhat.jenkins.nodesharing.utils.NodeSharingJenkinsRule.BlockingTask;
+import com.redhat.jenkins.nodesharing.utils.NodeSharingJenkinsRule.MockTask;
 import com.redhat.jenkins.nodesharingbackend.Api;
 import com.redhat.jenkins.nodesharingbackend.Pool;
 import com.redhat.jenkins.nodesharingbackend.Pool.Updater;
@@ -35,14 +35,11 @@ import com.redhat.jenkins.nodesharingbackend.ShareableComputer;
 import com.redhat.jenkins.nodesharingbackend.ShareableNode;
 import com.redhat.jenkins.nodesharingfrontend.SharedNodeCloud;
 import hudson.AbortException;
-import hudson.ExtensionList;
 import hudson.FilePath;
 import hudson.model.Computer;
-import hudson.model.FreeStyleBuild;
 import hudson.model.Label;
 import hudson.model.Node;
 import hudson.model.Queue;
-import hudson.model.queue.QueueTaskFuture;
 import hudson.model.queue.ScheduleResult;
 import hudson.slaves.DumbSlave;
 import jenkins.model.Jenkins;
