+++ jth-tests/src/test/java/com/redhat/jenkins/nodesharing/ReservationTaskTest.java
@@ -25,6 +25,7 @@ package com.redhat.jenkins.nodesharing;
 
 import com.gargoylesoftware.htmlunit.Page;
 import com.redhat.jenkins.nodesharing.utils.BlockingBuilder;
+import com.redhat.jenkins.nodesharing.utils.NodeSharingJenkinsRule;
 import com.redhat.jenkins.nodesharingbackend.Pool;
 import com.redhat.jenkins.nodesharingbackend.ReservationTask;
 import org.junit.Rule;
